@@ -1,24 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:zac/src/flutter/widgets/scroll_controller.dart';
-import 'package:zac/src/zac/shared_value.dart';
 
 import '../helper.dart';
 
 void main() {
-  testWidgets('Provide ScrollController', (tester) async {
-    await testWithContextsWraped(
+  testWidgets('ScrollController', (tester) async {
+    await testBuilder<ScrollController>(
       tester,
-      (child) => FlutterScrollController(
-        child: child,
-      ),
-      (getContext, getZacContext) {
-        expect(
-            ConsumeSharedValue<ScrollController>(
-                    family: FlutterScrollController.familyName)
-                .build(getContext(), getZacContext()),
-            isA<ScrollController>());
-      },
+      'f:1:ScrollController',
+      matcher: (matcher) => matcher,
     );
   });
 }
