@@ -4,7 +4,7 @@ import 'package:zac/src/flutter/all.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:zac/src/zac/action.dart';
-import 'package:zac/src/zac/shared_value.dart';
+import 'package:zac/src/zac/shared_state.dart';
 import 'package:zac/src/zac/widget.dart';
 import 'package:zac/src/zac/zac_value.dart';
 
@@ -250,15 +250,16 @@ void main() {
             data: FlutterMaterialApp(
               home: FlutterMaterial(
                 child: FlutterScaffold(
-                  body: SharedValueProviderBuilder.provideString(
-                    value: 'hello world',
-                    family: 'shared',
+                  body: ZacSharedStateProvider(
+                    states: {
+                      'shared': ZacSharedStateProvide.value('hello world'),
+                    },
                     child: FlutterElevatedButton(
                       key: FlutterValueKey('button'),
                       onPressed: ZacListOfActions([
                         FlutterDialogActions.showDialog(
                           child: FlutterText(
-                            ConsumeSharedValue<String>(family: 'shared'),
+                            SharedStateConsume<String>(family: 'shared'),
                           ),
                         ),
                       ]),
